requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
praw==7.7.1
prawcore==2.3.0
prompt-toolkit==3.0.36
Pygments==2.16.1
python-dotenv==1.0.0
questionary==2.0.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
tqdm==4.66.1
update-checker==0.18.0
urllib3==2.0.4
wcwidth==0.2.6
websocket-client==1.6.1
yarl==1.9.2
